+++ src/core/index.ts
@@ -1,2 +1,4 @@
 export { createShallowStore } from "./createShallowStore";
 export { createStoreToolkit } from "./createStoreToolkit";
+export type { ShallowStoreBindings } from "./createShallowStore.types";
+export type { StoreToolkit } from "./createStoreToolkit.types";