+++ src/hooks/storeSelection/index.ts
@@ -1,1 +1,2 @@
 export { useStoreSelection, useStoreSelectionPlain } from "./useStoreSelection";
+export type { StorePlainHook, StoreValueHook } from "./storeSelection.types";