+++ src/hooks/storeSelection/storeSelection.types.ts
@@ -4,3 +4,13 @@ export type SelectionStore<TState> = Pick<
   StoreApi<TState>,
   "getState" | "getInitialState" | "subscribe"
 >;
+
+export interface StoreValueHook<TState> {
+  (): TState;
+  <T>(selector: (state: TState) => T, equalityFn?: (left: T, right: T) => boolean): T;
+}
+
+export interface StorePlainHook<TState> {
+  (): TState;
+  <T>(selector: (state: TState) => T): T;
+}