+++ src/providers/index.ts
@@ -1,1 +1,6 @@
 export { createStoreProvider } from "./createStoreProvider";
+export type {
+  StoreProviderConfig,
+  StoreProviderProps,
+  StoreProviderResult,
+} from "./createStoreProvider.types";